+++ app/backend/src/main/java/org/letspeppol/app/dto/TotalProcessedDto.java
@@ -1,6 +1,6 @@
 package org.letspeppol.app.dto;
 
 public record TotalProcessedDto(
-    Integer totalProcessed,
-    Integer totalProcessedToday
+    Long totalProcessed,
+    Long totalProcessedToday
 ) {}
